did all but last function for project 3

+++ week_07/project_03/bank_account.hpp
@@ -12,7 +12,6 @@ void string_date_to_ints(std::string, int &, int &, int &);
 int number_of_first_of_months(std::string, std::string);
 double interest_earned(double, double, std::string, std::string);
 std::string process_command(std::string, std::string &, double &, double);
-
-std::string process_commands(std::string, int);
+std::string process_commands(std::string, double);
 
 #endif
